+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Version 3.1.1 *(In Development)*
 
 * Fix: `MenuItem.getSubMenu` now returns a support instance rather than a
   native instance.
+* Fix: Fragment methods `onAttach` and `onInflate` incorrectly regressed to use
+  `Activity` instead of a `FragmentActivity` in their method signatures.
 
 
 Version 3.1.0 *(2011-07-22)*