+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Version 3.1.1 *(In Development)*
   native instance.
 * Fix: Fragment methods `onAttach` and `onInflate` incorrectly regressed to use
   `Activity` instead of a `FragmentActivity` in their method signatures.
+* Fix: Retained fragments not being re-attached on pre-3.0 when attached to
+  `android.R.id.content` upon activity recreation.
 
 
 Version 3.1.0 *(2011-07-22)*